Update build.gradle, add IntelliJ run configs
I’ve updated the build.gradle to be formatted better, and add a BuildConstants file with build information. It also will format the code (which has also been done in this commit, which is why there are so many changes). I’ve also added IntelliJ run configs (not functional yet) to switch between real robot, simulation, and replay. There are still a couple of things that need to be finished up soon related to IntelliJ and Gradle.

Signed-off-by: Ashton A <uhdbits@duck.com>